Fix Memory Exhaust Issue
Issue Discription:

with an incoming http requests the server will allocate memory as long as it recieves Data from the
incoming Connection even if a MaxUploadSize was defined. This behaviour will trigger a bad_alloc exception
then the server host runs out of memory and the server application will terminate.
A example would be you upload a 1GB File via a POST request to a server with 512 MB of Memory and no swap.

Issue Fix:

append incoming Data to the content buffer only as long as its size is smaller than maxUploadFile.
Otherwise throw away the Data and than return a Status 403 page