Add Support for multiple Database Connections to CWF::SqlDatabaseStorage
with the old implementation there was a issue where you could really only
use one database with CWF::SqlDatabaseStorage in you're application.
This was due to the fact that CWF::SqlDatabaseStorage::getDatabase would always
return the first database connection created on the executing thread.
It would possibly ignore Configuration given to CWF::SqlDatabaseStorage which
a user might not expect. This patch fixes that.
It replaces the internal pointer to QSqldatabase with two maps. One maps the
application-database-connection names to unique-database-connection-names.
The second map maps unique-database-connection-names to pointers to QSqldatabase objects.
Also logic was added to to return the correct database-connection as well as
cleanup code when the thread would exit.